// src/common/enums.ts
export enum ReflectKeys {
  MODULE_KEY = 'module_key',
  METHOD_KEY = 'method_key',
  REQUEST_BODY_KEY = 'request_body_key',
  VALIDATE_KEY = 'validate_key',
  SELF_PARAM_KEY = 'design:paramtypes',
  LOGIN_KEY = 'login_key',
  PERMISSION_KEY = 'permission_key',
  DATABASE_KEY = 'database_key',
  TRANSACTION_KEY = 'transaction_key',
  ID_KEY = 'id_key'
}

export enum LoginType {
  ANONYMOUS = 'anonymous',
  ADMIN = 'admin',
  GROUP = 'group'
}